Move tests for np.linspace and np.zeros_like to their own test module
Also a small touch-up to the definitions in function_helpers.
